+++ resdk/resources/base.py
@@ -27,12 +27,15 @@ class BaseResource(object):
 
     endpoint = None
 
+    WRITABLE_FIELDS = ('slug', 'name', 'permissions')
+    UPDATE_PROTECTED_FIELDS = ('contributor', )
+    READ_ONLY_FIELDS = ('id', 'created', 'modified')
+
     def __init__(self, slug=None, id=None,  # pylint: disable=redefined-builtin
                  model_data=None, resolwe=None):
-        """Verify that one and only one of slug, id, model_data is given."""
-        identifiers = iter((slug, id, model_data))
-        if not any(identifiers) or any(identifiers):
-            raise ValueError("One and only one of slug, id or model_data must be given")
+        """Verify that only a single attribute of slug, id or model_data given."""
+        if len([attr for attr in (slug, id, model_data) if attr is not None]) > 1:
+            raise ValueError("Only one of slug, id or model_data allowed")
 
         if id:
             if not isinstance(id, int):
@@ -40,9 +43,14 @@ class BaseResource(object):
         elif slug:
             if not isinstance(slug, six.string_types):
                 raise ValueError("slug should be a string.")
-        else:
+        elif model_data:
             if not isinstance(model_data, dict):
                 raise ValueError("model_data should be a dict.")
+
+        self._original_values = {}
+
+        for field_name in self.fields():
+            setattr(self, field_name, None)
 
         #: a descriptive name of the resource
         self.name = None
@@ -50,16 +58,12 @@ class BaseResource(object):
         self.id = id  # pylint: disable=invalid-name
         #: human-readable unique identifier
         self.slug = slug
-
         #: user id of the contributor
         self.contributor = None
-
         #: date of creation
         self.created = None
-
         #: date of latest modification
         self.modified = None
-
         #: permissions - (view/download/add/edit/share/owner for user/group/public)
         self.permissions = None
 
@@ -88,19 +92,52 @@ class BaseResource(object):
 
         self.logger = logging.getLogger(__name__)
 
-    def _update_fields(self, fields):
+    def fields(self):
+        """Resource fields."""
+        return self.WRITABLE_FIELDS + self.UPDATE_PROTECTED_FIELDS + self.READ_ONLY_FIELDS
+
+    def _update_fields(self, payload):
         """Update resource fields.
 
-        :param fields: Resource fields
-        :type fields: dict
+        :param payload: Resource fields
+        :type payload: dict
 
         """
-        for field_name, field_value in fields.items():
-            setattr(self, field_name, field_value)
+        self._original_values = payload
+        for field_name in self.fields():
+            setattr(self, field_name, payload.get(field_name, None))
 
     def update(self):
         """Update resource fields from the server."""
-        self._update_fields(self.api(self.id).get())
+        response = self.api(self.id).get()
+        self._update_fields(response)
+
+    def save(self):
+        """Save resource to the server."""
+        if self.id:  # update resource
+            payload = {}
+            for field_name in self.WRITABLE_FIELDS:
+                # find changes
+                if getattr(self, field_name) != self._original_values.get(field_name):
+                    payload[field_name] = getattr(self, field_name)
+
+            if payload:
+                response = self.api(self.id).patch(payload)
+                self._update_fields(response)
+
+        else:  # create resource
+            try:
+                field_names = self.WRITABLE_FIELDS + self.UPDATE_PROTECTED_FIELDS
+                payload = {field_name: getattr(self, field_name) for field_name in field_names
+                           if getattr(self, field_name) is not None}
+
+                response = self.api.post(payload)
+                self._update_fields(response)
+            except slumber.exceptions.HttpClientError as ex:
+                # pylint: disable=no-member
+                raise slumber.exceptions.HttpClientError('{}\n\n{}'.format(ex.message, ex.content),
+                                                         response=ex.response,
+                                                         content=ex.content)
 
     def delete(self):
         """Delete the resource object from the server."""