+++ resdk/resources/process.py
@@ -24,9 +24,15 @@ class Process(BaseResource):
 
     endpoint = "process"
 
+    WRITABLE_FIELDS = ('data_name', 'version', 'type', 'flow_collection', 'category',
+                       'persistence', 'priority', 'description', 'input_schema',
+                       'output_schema', 'run') + BaseResource.WRITABLE_FIELDS
+
     def __init__(self, slug=None, id=None,  # pylint: disable=redefined-builtin
                  model_data=None, resolwe=None):
         """Initialize attributes."""
+        BaseResource.__init__(self, slug, id, model_data, resolwe)
+
         self.data_name = None
         """
         the default name of data object using this process. When data object
@@ -34,20 +40,15 @@ class Process(BaseResource):
         data object is determined from this field. The field is a expression
         which can take values of other fields.
         """
-
         #: the process version
         self.version = None
-
         #: the type of process ``"type:sub_type:sub_sub_type:..."``
         self.type = None
-
         #: Current options: "sample"/None. If sample, new "sample" will be created
         #: (if not already existing) and annotated with provided descriptor.
         self.flow_collection = None
-
         #: used to group processes in a GUI. Examples: ``upload:``, ``analyses:variants:``, ...
         self.category = None
-
         self.persistence = None
         """
         Measure of how important is to keep the process outputs when
@@ -58,23 +59,16 @@ class Process(BaseResource):
         analysis use CACHED - the results can stil be calculated from
         imported data but it can take time.
         """
-
         #: process priority - not used yet
         self.priority = None
-
         #: process description
         self.description = None
-
         #: specifications of inputs
         self.input_schema = None
-
         #: specification of outputs
         self.output_schema = None
-
         #: the heart of process - here the algorithm is defined.
         self.run = None
-
-        BaseResource.__init__(self, slug, id, model_data, resolwe)
 
     def print_inputs(self):
         """Pretty print input_schema."""